+++ Tests/Functions/Public/Set-JiraVersion.Unit.Tests.ps1
@@ -102,7 +102,7 @@ InModuleScope JiraPS {
                     @{ parameter = 'Released'; type = 'Boolean' }
                     @{ parameter = 'ReleaseDate'; type = 'DateTime' }
                     @{ parameter = 'StartDate'; type = 'DateTime' }
-                    @{ parameter = 'Project'; type = 'Object' }
+                    @{ parameter = 'Project'; type = 'AtlassianPS.JiraPS.Project' }
                     @{ parameter = 'Credential'; type = 'PSCredential' }
                 ) {
                     param($parameter, $type)